# frontend/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KYC API",
    "version": "1.0.0",
    "description": "JSON:API endpoints for Applicant accounts and sessions, KYC applications and document evidence, reviewer workload, and access checks. Authenticated endpoints use the first-party session cookie. Error responses use a JSON:API errors array with safe, client-actionable details."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current frontend origin; /api/v1 is proxied to the Java API in development."
    }
  ],
...
  "paths": {
...
    "/api/v1/reviewer-applications": {
      "get": {
        "operationId": "listReviewerApplications",
        "summary": "Read reviewer workload summary and a page of applications",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Case-insensitive applicant-name or application-ID query.",
            "schema": {
              "type": "string",
              "maxLength": 120
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Lifecycle status filter; omit or use all for all statuses.",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "draft",
                "submitted",
                "in-review",
                "approved",
                "rejected"
              ],
              "default": "all"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "page-size",
            "in": "query",
            "description": "Requested page size, from 1 to 50. Defaults to 10.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workload summary and matching page, newest first.",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewerApplicationsDocument"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ReviewerQueryInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          }
        },
        "tags": ["Reviewer applications"],
        "description": "Returns aggregate workload counts and a newest-first page of minimized application records. Search matches applicant names or application IDs without case sensitivity. The response includes the effective page, page size, total item count, and total pages."
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-KYCSESSION",
        "description": "Opaque, secure, HttpOnly cookie set after successful sign-in. The browser sends it automatically to the API origin."
      }
    },
    "schemas": {
...
      "ErrorDocument": {
        "type": "object",
        "required": ["errors"],
        "properties": {
          "errors": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "description": "HTTP status code represented as a string, such as \"422\"."
                },
                "code": {
                  "type": "string",
                  "description": "Stable machine-readable error code for client handling."
                },
                "title": {
                  "type": "string",
                  "description": "Short, human-readable summary of the error."
                },
                "detail": {
                  "type": "string",
                  "description": "Safe, actionable explanation. It does not contain credentials, internal exception details, or stored document contents."
                },
                "source": {
                  "type": "object",
                  "properties": {
                    "pointer": {
                      "type": "string",
                      "description": "JSON Pointer to the invalid member in the request document, such as /data/attributes/email."
                    }
                  },
                  "description": "Location of the invalid request member. Omitted when the error is not tied to one field.",
                  "required": ["pointer"]
                }
              },
              "required": ["status", "code", "title", "detail"],
              "description": "Safe JSON:API error details. source is present for field validation errors and points to the invalid request member."
            },
            "description": "One or more error objects. Each object describes one failure; validation can return multiple errors in the same response."
          }
        },
        "description": "JSON:API error document. The errors array contains one or more safe error objects; credential validation can return several field errors."
      },
...
      "ReviewerApplicationsDocument": {
        "type": "object",
        "required": ["data", "meta"],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewerApplicationResource"
            }
          },
          "meta": {
            "type": "object",
            "required": ["summary", "pagination"],
            "properties": {
              "summary": {
                "type": "object",
                "required": [
                  "pendingReview",
                  "inProgress",
                  "approvedThisMonth"
                ],
                "properties": {
                  "pendingReview": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "inProgress": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "approvedThisMonth": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "monthStart": {
                    "type": "string",
                    "format": "date-time",
                    "description": "UTC start of current calendar month"
                  },
                  "monthEndExclusive": {
                    "type": "string",
                    "format": "date-time",
                    "description": "UTC start of next calendar month"
                  },
                  "timeZone": {
                    "type": "string",
                    "const": "UTC"
                  }
                },
                "description": "Counts for pending review, in-progress applications, and approvals in the current UTC month."
              },
              "pagination": {
                "type": "object",
                "required": ["page", "pageSize", "totalItems", "totalPages"],
                "properties": {
                  "page": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "pageSize": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "totalItems": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "totalPages": {
                    "type": "integer",
                    "minimum": 0
                  }
                },
                "description": "Zero-based page details for this response."
              }
            },
            "description": "Aggregate workload counts and pagination details for the returned data page."
          }
        },
        "description": "JSON:API response containing a reviewer workload summary, one application page, and pagination metadata."
      },
      "ReviewerApplicationResource": {
        "type": "object",
        "required": ["type", "id", "attributes"],
        "properties": {
          "type": {
            "const": "reviewer-applications"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "attributes": {
            "type": "object",
            "required": ["applicantName", "status"],
            "properties": {
              "applicantName": {
                "type": "string"
              },
              "submittedAt": {
                "type": ["string", "null"],
                "format": "date-time",
                "description": "Submission time in UTC; null when the application has not been submitted."
              },
              "status": {
                "type": "string",
                "enum": [
                  "draft",
                  "submitted",
                  "in-review",
                  "approved",
                  "rejected"
                ]
              }
            }
          }
        },
        "description": "Minimized reviewer list resource. It contains applicant name and lifecycle status, not form answers or document data."
      }
    },
    "responses": {
...
      "AuthenticationRequired": {
        "description": "No active Applicant session is available. Sign in and send the session cookie.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "authenticationRequired": {
                "summary": "Session required",
                "value": {
                  "errors": [
                    {
                      "status": "401",
                      "code": "authentication-required",
                      "title": "Authentication required",
                      "detail": "Sign in to continue."
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "AccessDenied": {
        "description": "The request is not authorized for this resource or is missing a valid CSRF token. The response does not disclose protected resource details.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "accessDenied": {
                "summary": "Access denied",
                "value": {
                  "errors": [
                    {
                      "status": "403",
                      "code": "access-denied",
                      "title": "Access denied",
                      "detail": "You do not have permission to access this resource."
                    }
                  ]
                }
              }
            }
          }
        }
      },
...
      "NotAcceptable": {
        "description": "The requested response representation is not supported. Accept application/vnd.api+json.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "notAcceptable": {
                "summary": "Unsupported response representation",
                "value": {
                  "errors": [
                    {
                      "status": "406",
                      "code": "not-acceptable",
                      "title": "Not acceptable",
                      "detail": "Accept application/vnd.api+json for JSON:API responses."
                    }
                  ]
                }
              }
            }
          }
        }
      },
...
      "ReviewerQueryInvalid": {
        "description": "One or more reviewer query parameters are invalid. Check search, status, page, and page-size values.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "invalidQuery": {
                "summary": "Invalid reviewer query",
                "value": {
                  "errors": [
                    {
                      "status": "400",
                      "code": "invalid-query",
                      "title": "Invalid query",
                      "detail": "Check search, status, page, and page size."
                    }
                  ]
                }
              }
            }
          }
        }
      },
...
    },
...
  }
}